Refactor controller code
