+++ dice-server.js
@@ -1,14 +1,7 @@
-const express = require('express');
 const nconf = require('nconf');
-const bodyParser = require('body-parser');
 
-const app = express();
-const routerParams = { caseSensitive: true, strict: true };
-const rootRouter = express.Router(routerParams);
-
-
+// TODO move to dedicated module
 nconf.argv().env().file({ file: './config.json' });
-
 nconf.defaults({
   server: {
     protocol: 'http',
@@ -24,16 +17,13 @@ nconf.defaults({
     database: 'dicedb',
   },
 });
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-
 nconf.required(['server', 'database', 'smtp', 'private-key', 'public-key']);
 
+// TODO move imports to top once possible
 const { setupDb } = require('./src/api/db-handler');
 
 setupDb();
 
-require('./src/controller')(rootRouter, express.Router(routerParams));
+const controller = require('./src/controller');
 
-app.use(nconf.get('server:baseurl'), rootRouter);
-app.listen(nconf.get('server:port'), () => console.info(`Running on port ${nconf.get('server:port')}`));
+controller(nconf.get('server:baseurl'), nconf.get('server:port'));
