+++ src/api/validator.js
@@ -4,6 +4,7 @@ const crypto = require('crypto');
 
 const algorithm = 'RSA-SHA512';
 const encoding = 'base64';
+// TODO initialize elsewhere
 const privateKey = fs.readFileSync(nconf.get('private-key'));
 const publicKey = fs.readFileSync(nconf.get('public-key'));
